Refactor game class.
Refactored game class as TSE requested.
